gowin: add Video_PLL base class
